Revert "refactor code"
