+++ src/components/app/nfts/components/create-form.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 "use client"
 
 import { zodResolver } from "@hookform/resolvers/zod"
