+++ use-cardano/contexts/CardanoContext.tsx
@@ -1,9 +1,10 @@
 
 "use client";
 import React  from "react"
-import { DefaultUseCardanoOptions, UseCardanoContextState } from "../types";
+import {
+     DefaultUseCardanoOptions, UseCardanoContextState,
+} from "use-cardano"
 
-// eslint-disable-next-line no-unused-vars
 const noop = (..._: any[]) => {}
 
 export const defaultContextState: UseCardanoContextState = {
