+++ use-cardano/hooks/use-cardano-initialization.ts
@@ -1,0 +1,8 @@
+import { useLucidAndWalletApi } from "../hooks/use-lucid-and-wallet-api"
+import { useWalletProviders } from "../hooks/use-wallet-providers"
+import { UseCardanoOptionsWithDefaults } from "use-cardano"
+
+export const useCardanoInitialization = (options: UseCardanoOptionsWithDefaults) => {
+  useWalletProviders(options)
+  useLucidAndWalletApi(options)
+}
