+++ use-cardano/hooks/use-lucid-and-wallet-api.ts
@@ -1,4 +1,3 @@
-/* eslint-disable react-hooks/exhaustive-deps */
 import { useCardano } from "../contexts/CardanoContext"
 import {
     disallowedNetworkError, invalidWalletError, noAccountSetError, noDappError, unknownError,
@@ -15,7 +14,7 @@ import { noLiveNetworkChangeWarning } from "../lib/warnings"
 import { isNil } from "lodash"
 import { Lucid, Network, WalletApi } from "lucid-cardano"
 import { useCallback, useEffect } from "react"
-import { UseCardanoOptionsWithDefaults } from "../types"
+import { UseCardanoOptionsWithDefaults } from "use-cardano"
 
 export const useLucidAndWalletApi = (options: UseCardanoOptionsWithDefaults) => {
   const { autoReconnect, allowedNetworks: allowedNetworkNames, testnetNetwork, node } = options
