+++ use-cardano/hooks/use-wallet-providers.tsx
@@ -2,7 +2,7 @@ import { useCardano } from "../contexts/CardanoContext"
 import { filterAvailableProviders } from "../lib/filter-available-providers"
 import { getStoredWalletProvider, setStoredWalletProvider } from "../lib/local-storage"
 import { useEffect, useRef } from "react"
-import { UseCardanoOptionsWithDefaults } from "../types"
+import { UseCardanoOptionsWithDefaults } from "use-cardano"
 
 type Interval = ReturnType<typeof setInterval>
 
@@ -59,7 +59,6 @@ const useWalletProviders = (options: UseCardanoOptionsWithDefaults) => {
     return () => {
       if (interval.current) clearInterval(interval.current)
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 }
