+++ use-cardano/lib/get-toaster-texts.ts
@@ -1,4 +1,4 @@
-import { WalletProvider } from "../types"
+import { WalletProvider } from "use-cardano"
 
 export const getText = (provider: WalletProvider) => `Connected to ${provider}`
