+++ use-cardano/lib/supported-wallet-providers.ts
@@ -1,3 +1,3 @@
-import { WalletProvider } from "../types";
+import { WalletProvider } from "use-cardano"
 
 export const supportedWalletProviders: WalletProvider[] = ["nami", "eternl", "gerowallet", "flint"]
