+++ use-cardano/lib/utils/has-error-code.ts
