+++ use-cardano/lib/utils/is-error.ts
